+++ src/test/java/de/mdelab/predictor/loader/test/TestSaturatingModel.java
@@ -42,7 +42,7 @@ public class TestSaturatingModel {
 				
 				Double actual = 205.4556905;
 				
-				Float predicted = lrm.pointPrediction(userArguments);
+				Double predicted = lrm.pointPrediction(userArguments);
 				System.out.println("Predicted="+predicted+", Actual="+actual);
 
 				Double min = actual - actual*allowedPercentDeviation/100;
